splitting up large file
