+++ spyro/plots/__init__.py
@@ -1,5 +1,7 @@
-from .plots import plot_shots, plot_mesh_sizes, plot_model, plot_function, plot_model_in_p1
-from .plots import debug_plot, debug_pvd, plot_receiver_response
+from .general_plots import plot_shots, plot_function, plot_model, plot_model_in_p1
+from .mesh_plots import plot_mesh_sizes
+from .debug_plots import debug_plot, debug_pvd
+from .receiver_plots import plot_receiver_response
 
 __all__ = [
     "plot_shots",
@@ -10,4 +12,5 @@ __all__ = [
     "debug_pvd",
     "plot_model_in_p1",
     "plot_receiver_response",
+    "plot_displacement_components",
 ]
